requirements.txt
#framework
django==3.0.5
djangorestframework==3.10.3
django-cors-headers==3.1.1
django-widget-tweaks==1.4.8
drf-yasg==1.17.1
django-oauth-toolkit==1.3.2

#server
uwsgi==2.0.19.1

#data
psycopg2-binary==2.8.6
